Cast to int in loop

+++ run-length-encoding/src/run_length_encoding.c
@@ -10,7 +10,7 @@ char *encode(const char *text)
 {
     char *encoded = (char *)calloc(MAX_STRING_LEN, sizeof(char));
     int num_of_chars = 0;
-    for (size_t i = 0; i < strlen(text) - 1; i++)
+    for (int i = 0; i < (int)strlen(text) - 1; i++)
     {
         if (text[i] != text[i + 1])
         {
